+++ apps/gizual-app/src/types/visualization.ts
@@ -4,4 +4,7 @@ export type VisualisationConfig = {
     newest: string;
     notLoaded: string;
   };
+  style: {
+    lineLength: "lineLength" | "full";
+  };
 };